requirements-combined.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
requests==2.31.0
python-multipart==0.0.6
streamlit==1.42.0
streamlit-autorefresh==1.0.1
pandas
feedparser
tweepy
python-dateutil
geotext
pytz
schedule
pydeck
plotly
numpy
pytest==8.3.5
tenacity>=8.2.3
vader-sentiment==3.2.1
nltk>=3.8.1